requirements.txt
openai>=1.0.0
python-dotenv
pyautogen==0.2.8
numpy
fastapi
uvicorn
